ContestTelegramBot/requirements.txt
python-telegram-bot==12.7
pymysql
emoji~=0.6.0
mysql~=0.0.2
mysql-connector-python~=8.0.22